# services/core/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: FeastCloud Core API
  version: 0.1.0
  description: Phase 0 tenant-safe organization, kitchen, order, audit, and edge-inbox API.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/tenders/{id}/reverse:
    post: { tags: [Commerce], operationId: reverseTender, parameters: [{ $ref: "#/components/parameters/ResourceID" }, { $ref: "#/components/parameters/IdempotencyKey" }], requestBody: { $ref: "#/components/requestBodies/Mutation" }, responses: { "201": { $ref: "#/components/responses/Created" }, default: { $ref: "#/components/responses/Problem" } } }
...
  /api/v1/guests:
    get: { tags: [Guest], operationId: listGuests, responses: { "200": { $ref: "#/components/responses/Collection" }, default: { $ref: "#/components/responses/Problem" } } }
    post: { tags: [Guest], operationId: createGuest, parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }], requestBody: { $ref: "#/components/requestBodies/Mutation" }, responses: { "201": { $ref: "#/components/responses/Created" }, default: { $ref: "#/components/responses/Problem" } } }
  /api/v1/guests/{id}/consent:
    post: { tags: [Guest], operationId: setGuestConsent, parameters: [{ $ref: "#/components/parameters/ResourceID" }, { $ref: "#/components/parameters/IdempotencyKey" }], requestBody: { $ref: "#/components/requestBodies/Mutation" }, responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Problem" } } }
  /api/v1/reservations:
    get: { tags: [Guest], operationId: listReservations, parameters: [{ $ref: "#/components/parameters/OutletFilter" }], responses: { "200": { $ref: "#/components/responses/Collection" }, default: { $ref: "#/components/responses/Problem" } } }
    post: { tags: [Guest], operationId: createReservation, parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }], requestBody: { $ref: "#/components/requestBodies/Mutation" }, responses: { "201": { $ref: "#/components/responses/Created" }, default: { $ref: "#/components/responses/Problem" } } }
  /api/v1/reservations/{id}/transitions:
    post: { tags: [Guest], operationId: transitionReservation, parameters: [{ $ref: "#/components/parameters/ResourceID" }, { $ref: "#/components/parameters/IdempotencyKey" }], requestBody: { $ref: "#/components/requestBodies/Mutation" }, responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Problem" } } }
  /api/v1/promotions:
    get: { tags: [Guest], operationId: listPromotions, parameters: [{ $ref: "#/components/parameters/OutletFilter" }], responses: { "200": { $ref: "#/components/responses/Collection" }, default: { $ref: "#/components/responses/Problem" } } }
    post: { tags: [Guest], operationId: createPromotion, parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }], requestBody: { $ref: "#/components/requestBodies/Mutation" }, responses: { "201": { $ref: "#/components/responses/Created" }, default: { $ref: "#/components/responses/Problem" } } }
  /api/v1/promotions/redeem:
    post: { tags: [Guest], operationId: redeemPromotion, parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }], requestBody: { $ref: "#/components/requestBodies/Mutation" }, responses: { "201": { $ref: "#/components/responses/Created" }, default: { $ref: "#/components/responses/Problem" } } }
...
  /api/v1/loyalty-accounts/{id}/events:
    post: { tags: [Guest], operationId: adjustLoyalty, parameters: [{ $ref: "#/components/parameters/ResourceID" }, { $ref: "#/components/parameters/IdempotencyKey" }], requestBody: { $ref: "#/components/requestBodies/Mutation" }, responses: { "201": { $ref: "#/components/responses/Created" }, default: { $ref: "#/components/responses/Problem" } } }
...
security:
  - OIDC: []
components:
  securitySchemes:
    OIDC:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/.well-known/openid-configuration
      description: Production mode validates RS256 issuer, audience, tenant, role, outlet, and lifetime claims.
    OutletDeviceCertificate:
      type: mutualTLS
      description: Enrolled outlet device certificate; revocation is checked in PostgreSQL on each sync request.
    DemoTenant:
      type: apiKey
      in: header
      name: X-FeastCloud-Tenant-ID
      description: Phase 0 adapter only; production uses OIDC and device certificates.
    DemoActor:
      type: apiKey
      in: header
      name: X-FeastCloud-Actor-ID
      description: Phase 0 adapter only; production derives the actor from authentication.
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    OutletFilter:
      name: outletId
      in: query
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 256 }
...
  requestBodies:
    Mutation:
      required: true
      content:
        application/json:
          schema:
              $ref: "#/components/schemas/MutationEnvelope"
...
  responses:
    Created:
      description: Resource created. Replays preserve the business result and set Idempotency-Replayed.
      headers:
        Location:
          schema: { type: string }
        Idempotency-Replayed:
          schema: { type: boolean }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ResponseEnvelope" }
    Entity:
      description: A tenant-scoped entity.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ResponseEnvelope" }
    Collection:
      description: A cursor-paginated tenant-scoped collection.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/CollectionEnvelope" }
    Problem:
      description: Stable machine-readable problem details.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    MutationEnvelope:
      $ref: "../../../packages/contracts/schemas/mutation-envelope.json"
    ResponseEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta:
          type: object
          required: [requestId]
          properties:
            requestId: { type: string }
            idempotencyReplay: { type: boolean }
    CollectionEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: {} }
        meta:
          type: object
          required: [requestId, page]
          properties:
            requestId: { type: string }
            page:
              type: object
              required: [limit]
              properties:
                limit: { type: integer }
                nextCursor: { type: string }
    Problem:
      type: object
      required: [type, title, status, detail, code, messageKey, correlationId, retryable]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        messageKey: { type: string }
        correlationId: { type: string }
        retryable: { type: boolean }
        violations:
          type: array
          items:
            type: object
            required: [field, code, messageKey]
            properties:
              field: { type: string }
              code: { type: string }
              messageKey: { type: string }
...